refactor: introduce a common data module
