+++ src/lib.rs
@@ -1,3 +1,4 @@
+mod data;
 mod error;
 mod types;
 mod utils;
